Update ci.md
- Fix links
- Add links to outside documentation where additional background knowledge seems helpful
- Expanded a couple of points

Query: the hard-coded `tags: magland/labbox-ephys:0.4.10` in the build-and-push-image section, does this get updated too? Is the user responsible for updating it?
Query: Obviously we would like the tags of the version in the source files to match the version in the tag applied to the release, but is there any mechanism actually enforcing that these be the same?